@@ -1,10 +1,10 @@
-﻿using Microsoft.AspNetCore.Hosting;
+﻿using System.Collections.Generic;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
-using System.Collections.Generic;
-using System.Net.Http;
-using System.Threading.Tasks;
 using TestEnvironment.Docker;
 using TestEnvironment.Docker.Containers.Mongo;
 using Xunit;
@@ -14,21 +14,21 @@ namespace AspNetCoreSample.Tests
     public class EnvironmentFixture : IAsyncLifetime
     {
         private readonly string _environmentName;
-        private DockerEnvironment _environment;
+        private IDockerEnvironment _environment;
         private IHost _host;
 
         public HttpClient TestClient { get; private set; }
 
-        public EnvironmentFixture(string environmentName)
-        {
+#pragma warning disable SA1201 // Elements should appear in the correct order
+        public EnvironmentFixture(string environmentName) =>
+#pragma warning restore SA1201 // Elements should appear in the correct order
             _environmentName = environmentName;
-        }
 
         public async Task InitializeAsync()
         {
             // Docker environment seutup.
             _environment = CreateTestEnvironmentBuilder().Build();
-            await _environment.Up();
+            await _environment.UpAsync();
 
             // API Test host setup
             var mongoContainer = _environment.GetContainer<MongoContainer>("mongo");
@@ -38,7 +38,9 @@ namespace AspNetCoreSample.Tests
             await OnInitialized(mongoContainer);
         }
 
+#pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
         public async Task DisposeAsync()
+#pragma warning restore CS1998 // Async method lacks 'await' operators and will run synchronously
         {
             TestClient.Dispose();
             _host.Dispose();
@@ -54,15 +56,23 @@ namespace AspNetCoreSample.Tests
             new DockerEnvironmentBuilder()
                 .SetName(_environmentName)
 #if DEBUG
-                .AddMongoContainer("mongo", reuseContainer: true);
+                .AddMongoContainer(p => p with
+                {
+                    Name = "mongo",
+                    Reusable = true
+                });
 #else
-                .AddMongoContainer("mongo");
+                .AddMongoContainer(p => p with
+                {
+                    Name = "mongo"
+                });
 #endif
 
         private IHostBuilder CreateHostBuilder(MongoContainer mongoContainer)
         {
             var config = new ConfigurationBuilder()
-                .AddInMemoryCollection(new[] {
+                .AddInMemoryCollection(new[]
+                {
                     new KeyValuePair<string, string>("MongoDbConnectionString", mongoContainer.GetConnectionString())
                 });
 
