@@ -2,7 +2,8 @@
 {
     public class WriteEnvironmentFixture : EnvironmentFixture
     {
-        public WriteEnvironmentFixture() : base("env1")
+        public WriteEnvironmentFixture()
+            : base("env1")
         {
         }
     }
