@@ -1,6 +1,6 @@
+﻿using System;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
-using System;
 
 namespace AspNetCoreSample
 {
