@@ -1,35 +1,34 @@
 ﻿using System.Collections.Generic;
 using Docker.DotNet;
 using Microsoft.Extensions.Logging;
+using TestEnvironment.Docker.ContainerOperations;
+using TestEnvironment.Docker.ImageOperations;
 using IP = System.Net.IPAddress;
 
 namespace TestEnvironment.Docker.Containers.Kafka
 {
-    public class KafkaContainer : TestEnvironment.Docker.Container
+    public class KafkaContainer : Container
     {
-        public KafkaContainer(DockerClient dockerClient, string name, string imageName = "johnnypark/kafka-zookeeper", string tag = "latest", IDictionary<string, string> environmentVariables = null, IDictionary<ushort, ushort> ports = null, bool isDockerInDocker = false, bool reuseContainer = false, ILogger logger = null)
-            : base(
-                dockerClient,
-                name,
-                imageName,
-                tag,
-                new Dictionary<string, string>
-                {
-                    ["ADVERTISED_HOST"] = "localhost",
-                }.MergeDictionaries(environmentVariables),
-                new Dictionary<ushort, ushort>
-                {
-                    [9092] = 9092,
-                    [2181] = 2181,
-                }.MergeDictionaries(ports),
-                isDockerInDocker,
-                reuseContainer,
-                new KafkaContainerWaiter(logger),
-                new KafkaContainerCleaner(),
-                logger)
+        public KafkaContainer(ContainerParameters containerParameters)
+            : base(containerParameters)
         {
         }
 
-        public string GetUrl() => IsDockerInDocker ? $"{IPAddress}:9092" : $"{IP.Loopback}:{Ports[9092]}";
+        public KafkaContainer(ContainerParameters containerParameters, IDockerClient dockerClient)
+            : base(containerParameters, dockerClient)
+        {
+        }
+
+        public KafkaContainer(ContainerParameters containerParameters, IDockerClient dockerClient, ILogger? logger)
+            : base(containerParameters, dockerClient, logger)
+        {
+        }
+
+        public KafkaContainer(ContainerParameters containerParameters, IContainerApi containerApi, ImageApi imageApi, ILogger? logger)
+            : base(containerParameters, containerApi, imageApi, logger)
+        {
+        }
+
+        public string GetUrl() => IsDockerInDocker ? $"{IPAddress}:9092" : $"{IP.Loopback}:{Ports![9092]}";
     }
 }
