@@ -4,17 +4,22 @@ using System.Threading.Tasks;
 using Elasticsearch.Net;
 using Microsoft.Extensions.Logging;
 using Nest;
+using TestEnvironment.Docker.ContainerLifecycle;
 
 namespace TestEnvironment.Docker.Containers.Elasticsearch
 {
     public class ElasticsearchContainerWaiter : BaseContainerWaiter<ElasticsearchContainer>
     {
-        public ElasticsearchContainerWaiter(ILogger logger = null)
+        public ElasticsearchContainerWaiter()
+        {
+        }
+
+        public ElasticsearchContainerWaiter(ILogger logger)
             : base(logger)
         {
         }
 
-        protected override async Task<bool> PerformCheck(ElasticsearchContainer container, CancellationToken cancellationToken)
+        protected override async Task<bool> PerformCheckAsync(ElasticsearchContainer container, CancellationToken cancellationToken)
         {
             var elastic = new ElasticClient(new Uri(container.GetUrl()));
             var health = await elastic.Cluster.HealthAsync(
