@@ -3,49 +3,46 @@ using System.Threading;
 using System.Threading.Tasks;
 using FluentFTP;
 using Microsoft.Extensions.Logging;
+using TestEnvironment.Docker.ContainerLifecycle;
 
 namespace TestEnvironment.Docker.Containers.Ftp
 {
     public class FtpContainerCleaner : IContainerCleaner<FtpContainer>
     {
-        private readonly ILogger _logger;
+        private readonly ILogger? _logger;
 
-        public FtpContainerCleaner(ILogger logger = null)
+        public FtpContainerCleaner()
         {
-            _logger = logger;
         }
 
-        public async Task Cleanup(FtpContainer container, CancellationToken token = default)
-        {
-            if (container == null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
+        public FtpContainerCleaner(ILogger logger) =>
+            _logger = logger;
 
-            using (var ftpClient = new FtpClient(container.FtpHost, container.IsDockerInDocker ? 21 : container.Ports[21], container.FtpUserName, container.FtpPassword))
+        public async Task CleanupAsync(FtpContainer container, CancellationToken cancellationToken = default)
+        {
+            using var ftpClient = new FtpClient(container.FtpHost, container.IsDockerInDocker ? 21 : container.Ports![21], container.FtpUserName, container.FtpPassword);
+            try
             {
-                try
+                await ftpClient.ConnectAsync(cancellationToken);
+                foreach (var item in await ftpClient.GetListingAsync("/"))
                 {
-                    await ftpClient.ConnectAsync(token);
-                    foreach (var item in await ftpClient.GetListingAsync("/"))
+                    if (item.Type == FtpFileSystemObjectType.Directory)
                     {
-                        if (item.Type == FtpFileSystemObjectType.Directory)
-                        {
-                            await ftpClient.DeleteDirectoryAsync(item.FullName, token);
-                        }
-                        else if (item.Type == FtpFileSystemObjectType.File)
-                        {
-                            await ftpClient.DeleteFileAsync(item.FullName, token);
-                        }
+                        await ftpClient.DeleteDirectoryAsync(item.FullName, cancellationToken);
+                    }
+                    else if (item.Type == FtpFileSystemObjectType.File)
+                    {
+                        await ftpClient.DeleteFileAsync(item.FullName, cancellationToken);
                     }
                 }
-                catch (Exception e)
-                {
-                    _logger?.LogInformation($"Cleanup issue: {e.Message}");
-                }
+            }
+            catch (Exception e)
+            {
+                _logger?.LogInformation($"Cleanup issue: {e.Message}");
             }
         }
 
-        public Task Cleanup(Container container, CancellationToken token = default) => Cleanup((FtpContainer)container, token);
+        public Task CleanupAsync(Container container, CancellationToken cancellationToken = default) =>
+            CleanupAsync((FtpContainer)container, cancellationToken);
     }
 }
