@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using TestEnvironment.Docker.ContainerLifecycle;
 
 namespace TestEnvironment.Docker.Containers.Mail
 {
@@ -11,25 +12,19 @@ namespace TestEnvironment.Docker.Containers.Mail
     {
         private readonly ushort _apiPort;
         private readonly string _deleteEndpoint;
-        private readonly ILogger _logger;
+        private readonly ILogger? _logger;
 
-        public MailContainerCleaner(ushort apiPort = 8025, string deleteEndpoint = "api/v1/messages", ILogger logger = null)
+        public MailContainerCleaner(ushort apiPort = 8025, string deleteEndpoint = "api/v1/messages") =>
+            (_apiPort, _deleteEndpoint) = (apiPort, deleteEndpoint);
+
+        public MailContainerCleaner(ILogger logger, ushort apiPort = 8025, string deleteEndpoint = "api/v1/messages") =>
+            (_logger, _apiPort, _deleteEndpoint) = (logger, apiPort, deleteEndpoint);
+
+        public async Task CleanupAsync(MailContainer container, CancellationToken cancellationToken = default)
         {
-            _apiPort = apiPort;
-            _deleteEndpoint = deleteEndpoint;
-            _logger = logger;
-        }
-
-        public async Task Cleanup(MailContainer container, CancellationToken token = default)
-        {
-            if (container == null)
-            {
-                throw new ArgumentNullException(nameof(container));
-            }
-
             var uri = new Uri($"http://" +
                 $"{(container.IsDockerInDocker ? container.IPAddress : IPAddress.Loopback.ToString())}:" +
-                $"{(container.IsDockerInDocker ? _apiPort : container.Ports[_apiPort])}");
+                $"{(container.IsDockerInDocker ? _apiPort : container.Ports![_apiPort])}");
 
             using (var httpClient = new HttpClient { BaseAddress = uri })
             {
@@ -49,6 +44,7 @@ namespace TestEnvironment.Docker.Containers.Mail
             }
         }
 
-        public Task Cleanup(Container container, CancellationToken token = default) => Cleanup((MailContainer)container, token);
+        public Task CleanupAsync(Container container, CancellationToken cancellationToken = default) =>
+            CleanupAsync((MailContainer)container, cancellationToken);
     }
 }
