@@ -1,43 +1,40 @@
 ﻿using System.Collections.Generic;
 using Docker.DotNet;
 using Microsoft.Extensions.Logging;
+using TestEnvironment.Docker.ContainerOperations;
+using TestEnvironment.Docker.ImageOperations;
 using IP = System.Net.IPAddress;
 
 namespace TestEnvironment.Docker.Containers.Mongo
 {
     public class MongoContainer : Container, IMongoContainer
     {
-        private readonly string _userName;
-        private readonly string _userPassword;
+        private readonly MongoContainerParameters _parameters;
 
-        public MongoContainer(
-            DockerClient dockerClient,
-            string name,
-            string userName,
-            string userPassword,
-            string imageName,
-            string tag = "latest",
-            IDictionary<string, string> environmentVariables = null,
-            IDictionary<ushort, ushort> ports = null,
-            bool isDockerInDocker = false,
-            bool reuseContainer = false,
-            IContainerWaiter containerWaiter = null,
-            IContainerCleaner containerCleaner = null,
-            ILogger logger = null)
-            : base(dockerClient, name, imageName, tag, environmentVariables, ports, isDockerInDocker, reuseContainer, containerWaiter, containerCleaner, logger)
-        {
-            _userName = userName;
-            _userPassword = userPassword;
-        }
+        public MongoContainer(MongoContainerParameters containerParameters)
+            : base(containerParameters) =>
+            _parameters = containerParameters;
+
+        public MongoContainer(MongoContainerParameters containerParameters, IDockerClient dockerClient)
+            : base(containerParameters, dockerClient) =>
+            _parameters = containerParameters;
+
+        public MongoContainer(MongoContainerParameters containerParameters, IDockerClient dockerClient, ILogger? logger)
+            : base(containerParameters, dockerClient, logger) =>
+            _parameters = containerParameters;
+
+        public MongoContainer(MongoContainerParameters containerParameters, IContainerApi containerApi, ImageApi imageApi, ILogger? logger)
+            : base(containerParameters, containerApi, imageApi, logger) =>
+            _parameters = containerParameters;
 
         public string GetConnectionString()
         {
             var hostname = IsDockerInDocker ? IPAddress : IP.Loopback.ToString();
-            var port = IsDockerInDocker ? 27017 : Ports[27017];
+            var port = IsDockerInDocker ? 27017 : Ports![27017];
 
-            return string.IsNullOrEmpty(_userName) || string.IsNullOrEmpty(_userPassword)
+            return string.IsNullOrEmpty(_parameters.UserName) || string.IsNullOrEmpty(_parameters.Password)
                 ? $@"mongodb://{hostname}:{port}"
-                : $@"mongodb://{_userName}:{_userPassword}@{hostname}:{port}";
+                : $@"mongodb://{_parameters.UserName}:{_parameters.Password}@{hostname}:{port}";
         }
     }
 }
