@@ -1,47 +1,37 @@
 ﻿using System.Collections.Generic;
 using Docker.DotNet;
 using Microsoft.Extensions.Logging;
+using TestEnvironment.Docker.ContainerOperations;
+using TestEnvironment.Docker.ImageOperations;
 using IP = System.Net.IPAddress;
 
 namespace TestEnvironment.Docker.Containers.Postgres
 {
     public class PostgresContainer : Container
     {
-        private readonly string _userName;
-        private readonly string _password;
+        private readonly PostgresContainerParameters _parameters;
 
-        public PostgresContainer(
-            DockerClient dockerClient,
-            string name,
-            string userName,
-            string password,
-            string imageName = "postgres",
-            string tag = "latest",
-            IDictionary<string, string> environmentVariables = null,
-            IDictionary<ushort, ushort> ports = null,
-            bool isDockerInDocker = false,
-            bool reuseContainer = false,
-            IContainerWaiter containerWaiter = null,
-            IContainerCleaner containerCleaner = null,
-            ILogger logger = null)
-            : base(
-                  dockerClient,
-                  name,
-                  imageName,
-                  tag,
-                  environmentVariables,
-                  ports,
-                  isDockerInDocker,
-                  reuseContainer,
-                  containerWaiter,
-                  containerCleaner,
-                  logger)
-        {
-            _userName = userName;
-            _password = password;
-        }
+        public string UserName => _parameters.UserName;
+
+#pragma warning disable SA1201 // Elements should appear in the correct order
+        public PostgresContainer(PostgresContainerParameters containerParameters)
+#pragma warning restore SA1201 // Elements should appear in the correct order
+            : base(containerParameters) =>
+            _parameters = containerParameters;
+
+        public PostgresContainer(PostgresContainerParameters containerParameters, IDockerClient dockerClient)
+            : base(containerParameters, dockerClient) =>
+            _parameters = containerParameters;
+
+        public PostgresContainer(PostgresContainerParameters containerParameters, IDockerClient dockerClient, ILogger? logger)
+            : base(containerParameters, dockerClient, logger) =>
+            _parameters = containerParameters;
+
+        public PostgresContainer(PostgresContainerParameters containerParameters, IContainerApi containerApi, ImageApi imageApi, ILogger? logger)
+            : base(containerParameters, containerApi, imageApi, logger) =>
+            _parameters = containerParameters;
 
         public string GetConnectionString() =>
-            $"Host={(IsDockerInDocker ? IPAddress : IP.Loopback.ToString())};Port={(IsDockerInDocker ? 5432 : Ports[5432])};Username={_userName};Password={_password}";
+            $"Host={(IsDockerInDocker ? IPAddress : IP.Loopback.ToString())};Port={(IsDockerInDocker ? 5432 : Ports![5432])};Username={_parameters.UserName};Password={_parameters.Password}";
     }
 }
