@@ -5,7 +5,8 @@ namespace TestEnvironment.Docker
 {
     public static class DictionaryExtensions
     {
-        public static IDictionary<T1, T2> MergeDictionaries<T1, T2>(this IDictionary<T1, T2> dictionary, IDictionary<T1, T2> other)
+        public static IDictionary<T1, T2> MergeDictionaries<T1, T2>(this IDictionary<T1, T2> dictionary, IDictionary<T1, T2>? other)
+            where T1 : notnull
         {
             if (other is null)
             {
