@@ -1,5 +1,5 @@
-﻿using System.Collections.Generic;
-using System.IO;
+﻿using System;
+using System.Collections.Generic;
 using Docker.DotNet;
 using Microsoft.Extensions.Logging;
 
@@ -7,34 +7,21 @@ namespace TestEnvironment.Docker
 {
     public interface IDockerEnvironmentBuilder
     {
-        DockerClient DockerClient { get; }
+        IDockerClient DockerClient { get; }
 
-        ILogger Logger { get; }
-
-        bool IsDockerInDocker { get; }
-
-        string EnvironmentName { get; }
+        ILogger? Logger { get; }
 
         IDockerEnvironmentBuilder DockerInDocker(bool dockerInDocker = true);
 
-        IDockerEnvironmentBuilder UseDefaultNetwork();
-
-        IDockerEnvironmentBuilder WithLogger(ILogger logger);
-
-        IDockerEnvironmentBuilder IgnoreFolders(params string[] ignoredFolders);
-
         IDockerEnvironmentBuilder SetName(string environmentName);
 
-        IDockerEnvironmentBuilder SetVariable(IDictionary<string, string> variables);
+        IDockerEnvironmentBuilder SetEnvironmentVariables(IDictionary<string, string> environmentVariables);
 
-        IDockerEnvironmentBuilder AddDependency(IDependency dependency);
+        IDockerEnvironmentBuilder AddContainer(Func<ContainerParameters, ContainerParameters> paramsBuilder);
 
-        IDockerEnvironmentBuilder AddContainer(string name, string imageName, string tag = "latest", IDictionary<string, string> environmentVariables = null, IDictionary<ushort, ushort> ports = null, bool reuseContainer = false, IContainerWaiter containerWaiter = null, IContainerCleaner containerCleaner = null);
+        IDockerEnvironmentBuilder AddContainer<TParams>(TParams containerParameters, Func<TParams, IDockerClient, ILogger?, Container> containerFactory)
+            where TParams : ContainerParameters;
 
-        IDockerEnvironmentBuilder AddFromCompose(Stream composeFileStream);
-
-        IDockerEnvironmentBuilder AddFromDockerfile(string name, string dockerfile, IDictionary<string, string> buildArgs = null, string context = ".", IDictionary<string, string> environmentVariables = null, IDictionary<ushort, ushort> ports = null, bool reuseContainer = false, IContainerWaiter containerWaiter = null, IContainerCleaner containerCleaner = null);
-
-        DockerEnvironment Build();
+        IDockerEnvironment Build();
     }
 }
